+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+<<<<<<< HEAD
 ## [2.0.3] 4th August 2025
 - Bug Fix for Web Popup Exit Intent 
 
 ## [2.0.2] 31st July 2025
 - Introduced preview in 'Advanced Web-popup Builder' in 'Web Popup'.
 
+=======
+>>>>>>> origin/fetch-api
 ## [2.0.1] 25th July 2025
 - Corrected click actions for web popup image only template
+- Added Fetch API support for network requests
 
 ## [2.0.0] 22nd July, 2025
 - Added Delivery Preference Support for Web Popups & Web Native Displays